修改计时溢出bug
全项目17个文件~40处 CInt(Timer) → CLng(Timer)，CInt仅±32767无法容纳Timer(0~86400)
修复XL算展取谕组中文路径429错误，Dir/MkDir→FileSystemObject

+++ 昭明计划VS优化_vba/WMR现市_ZZ基础.bas
@@ -913,7 +913,7 @@ Function 市调预程引擎初始(天组 As Variant, 天典 As Dictionary) As St
 '    Set WS花册 = Nothing
 '    MSG = MSG & "【更新花册名称】耗时：" & Round(Timer - TTT, 2) & vbCrLf
 '========================================================================================
-    MSG = MSG & "【A股行情数据爬取】耗时：" & CInt(Timer - TT) & vbCrLf
+    MSG = MSG & "【A股行情数据爬取】耗时：" & CLng(Timer - TT) & vbCrLf
     市调预程引擎初始 = MSG
 End Function
